move VpDirEntry to separate file VpFile.cs refactor constructor as ReadHeader
